updated tests and moved file to shared so I can use it for f2
