+++ src/applications/ivc-champva/10-7959f-1/helpers/CustomSSN.jsx
@@ -5,7 +5,7 @@ import { VaButton } from '@department-of-veterans-affairs/component-library/dist
 import SSNReviewWidget from 'platform/forms-system/src/js/review/SSNWidget';
 import get from 'platform/utilities/data/get';
 import { vaFileNumberUI } from 'platform/forms-system/src/js/web-component-patterns';
-import HandlePrefilledSSN, { maskSSN } from './maskSSN';
+import HandlePrefilledSSN, { maskSSN } from '../../shared/components/maskSSN';
 
 const SSN_DEFAULT_TITLE = 'Social Security number';
